Added correction - Change folder structure - Create impl for Services - Create a clase for Serilize - Added new dependency lombok - Change folder structure for test
