+++ src/main/java/com/shipping/application/util/MessageSerializer.java
@@ -1,0 +1,9 @@
+package com.shipping.application.util;
+
+public class MessageSerializer {
+
+    static public String requestMessageToPlainTextJson(RequestMessage requestMessage){
+            return String.format("{\"type\":\"%s\"}", requestMessage.type);
+    }
+
+}
