+++ src/main/java/com/shipping/application/util/RequestMessage.java
@@ -1,0 +1,15 @@
+package com.shipping.application.util;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+
+@AllArgsConstructor
+public class RequestMessage {
+
+    @Getter
+    @Setter
+    public String type;
+
+}
